requirements.txt
streamlit==1.10.0
python-docx==0.8.11
PyPDF2==3.0.0
langchain==0.0.137
langchain_community==0.0.4
transformers==4.18.0
pandas==1.4.2
python-dotenv==0.20.0
